fix hardcoded lists of machine types
the list of machine types is now generated by inspecting the subclasses of Machine, calling their device_type_name method to get a name string. this name string can be used in the device_type_by_name method of Machine to get the matching class to a string.

This removes the need to have an additional list / constants for the different types. Inheriting from Machine and implementing the device_type_name method adds a class to the list.