send the announcement as reply to machine owner
to have the announcement sent to the machines owner, first create a
rt ticket, then add a reply with the machines owner contact address
as cc.
the reason for this is that without additional configuration, created
tickets aren't send to the cc, and using the contact address as
requestor adds the usual "we've received your message" preamble,
which we don't want in this case.