WIP preview maintenance announcements
after creating an announcement, the generated tickets are
displayed for review. the announcement has an additional
attribute "preview" to facilitate this.

if approved, the tickets are created in rt, and the preview flag is
removed.

the case where the creation is aborted isn't well polished yet, it
creates an additional announcement from the previous hidden preview.