use rt queue from configuration instead of first owners
if multiple owners were involved in a single announcement _AND_
a value was set in the custom email address field, the ticket was
created in the wrong queue if the first owner had a custom queue
configured.

this fixes the queue selection so that if custom email addresses
are used the queue from the IDB configuration is used.