fix sending tickets
while sending tickets, it was tested with ".empty?" if an invitation should
be sent. the method whose value was tested with ".empty?" could also return
"nil", which doesn't have .empty?.

replace this ".empty?" with ".blank?", also return the empty string in the
respective method to prevent other places from tripping here.